benchmark/requirements.txt
psycopg2-binary
psycopg2
mysql-connector-python